Removed common endpoint from application secret auth code
The common endpoint does not support authentication with only clientId and secret. Because the application can exist in multiple tenants, there is no "home tenant" for the common endpoint to authenticate against. We have repro'd this in our tenant, and the fix was to remove /common/ and use the full Tenant ID, or the "tenant.onmicrosoft.com" equivalent

See also: https://github.com/MicrosoftDocs/powerapps-docs/issues/376