Remove duplicated sentences
The removed sentences 'Complex types are keyless named structured types consisting of a set of properties. They’re commonly used as property values in model entities, or as parameters or return values for operations.' are the exact same sentences used just one line up, at the introduction to complex types.