Create settings.md
Introducing a new page for canvas app settings content that doesn't seem to fit in any existing articles. App settings is getting a new control that requires documentation and it's expected to get a few more in the coming months, it may make sense to have this dedicated page for simplicity.