Update configure-openid-settings.md
Because implicit grant is not recommended by OAuth anymore. and some organisation might choose to support Hybrid Flow.  Also, Microsoft is working on support of authorisation flow grant (release date not confirmed yet).  I also added which "site setting" to updated based on experimentation, please confirm this is the correct way of doing it too.