Update configure-entity-relationship-cascading-behavior.md
Providing clarity for customer ask on tables and active state. 
Driven off customer discussions:
-------------------------------
Whether the codes are custom or out of box, almost every table that is considered Active for cascading is Statecode=0. This means if it’s one of the out of the box options, or something custom, if the enum value for statecode not statecode-0 it will not be included. This is driven by the enum values defined in our code.

Examples: 
•	Accounts: “Active” is active
•	Appointments: “Open” is active
•	Email: “Open” is active

There are a few exceptions where a statecode other than 0 is used (Quote, ServiceAppointment, Contract, RecurringAppointmentMaster).  But even in those cases, each table has been configured OOB to behave with a specific enum value. None of our tables use multiple statecode values as Active OOB.

“Scheduled” on Appointments is Statecode-3. To define this as active we would have to change the enumeration value for all users. A change to allow additional or alternate state codes to be considered “active” in our OOB configuration is something that would have global impact to every user of the system where these codes are used. Anyone who previously counted on Scheduled not being considered active would now be in an error condition

-------------------------------