Add section on how to choose a value for the FCC
Some customers are confused why the message "source record not selected" shows after they add a form component control. They don't know what to do to select a record; some expect the form component control to show a lookup control. So, there should be a section that mentions a way to select a record that references the message customers see when they first add a FCC.