Update connections-list.md
Adding docs for the Power Apps consent dialog - a UI experience that has existed since GA but hasn't ever had official docs. The intent with this docs is to address questions that have popped up periodically and answer questions for new enhancements coming to the consent dialog (e.g. when it can be suppressed).