More clarification on setting the Photo property
The **Photo** property is only set when the camera control is tapped or selected. Adding this to clarify as this is something that caused me wasted time trying to work out why I couldn't get an image when using a button, and has caused a number of questions across the internet.