Removed 'Crab Apples'
Removed 'Crab Apples' as it does not exist in the example/screenshots and causes confusion when reading. It states 'Crab apples' should exist (does not display in the first screenshot or second screenshot of the Examples section) and that it should not exist on the third. This causes the reader to think there's a copy and past error somewhere as 'Crab Apples' do not exist anywhere in this example.