Update Action Options. and made other amendments.
Updated the action choice options from 74 to 76 in line with https://learn.microsoft.com/en-us/power-apps/developer/data-platform/reference/entities/audit#action-choicesoptions

I made a few wording and grammar changes to make this translate better into other languages with web-based translators. For example changing 'At the time this topic was written' to 'When this topic was written'. I ensured that this did not change the meaning of the sentences in english but simply enabled them to translate slightly better.