Warn app makers about a common security pitfall for configuring forms
Many app makers rely on hiding or disabling form fields as a "secure" way to prevent users from reading or updating field values. This is not secure because it does not prevent users from accessing that data in other ways. Hiding and disabling HTML elements is meant as a usability feature, so that users will get validation errors earlier. It should not be used as a security feature.

For security, they should set up field level permissions on the backend.