Update file-column-data.md
It is important to architects designing systems(like business process management systems) which scale to millions of records to understand if the actual file is stored in the underlying SQL/relational storage or in a storage more optimized to hold files.  We allude to this in https://learn.microsoft.com/en-us/power-platform/admin/free-storage-space but its extremely difficult to arrive at this when someone is investigating the File column. 
There are a number of competitors who have a similar functionality which stores the file in a database. This will definitely be a differentiator in our favor.