Remove "Everyone" update share-app.md
I've made changes to remove mention of the "Everyone" group from this document. Microsoft is recommending that makers do not share their canvas apps with "Everyone" due to the overexposure that comes with sharing with that group. We are planning to eventually phase out the ability to share with "Everyone".