Update share-app.md
Simplified sharing experience for canvas apps. The scope of this is for only sharing canvas apps (Milestone 1). The other milestones will be released subsequently and will cover sharing model driven apps. 

Screenshots noted in the change will need to be replaced, and files will be shared to Manjinder / content team via email.