Update data-platform-powerbi-connector.md
The Error message: Return records size can't exceed 83886080 is obsolete, the limit of 80MB was removed a year ago. - There's a 5 minute time-out limit and this error message is never triggered. - 

Leaving it in place is misleading since it states that there's an 80MB limit, but that's no longer valid.