Update rename-data-source.md
Please verify if the "rename datasource" feature is available in GCC, GCC-H and DoD clouds and update the code change to reference the facts. Note: a lot of microsoft documentation online for power platform doesn't always state if the feature is available in the GCC, GCC-H, and DOD cloud environments.