Update aggregate-data.md
The Sample data is wrong for the date grouping so that needed to be fixed.
There was no information on how to access the returned values, which was very confusing to me.  Added details that they are returned as an Aliased value, and that the column will not be returned at all if it is a null grouping value.